+++ src/extensions/mvs/mvs-data.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2023 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2023-2024 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author Adam Midlik <midlik@gmail.com>
  */
@@ -9,27 +9,73 @@ import { treeToString } from './tree/generic/tree-utils';
 import { Root, createMVSBuilder } from './tree/mvs/mvs-builder';
 import { MVSTree, MVSTreeSchema } from './tree/mvs/mvs-tree';
 
-
-/** Top level of the MolViewSpec (MVS) data format. */
-export interface MVSData {
-    /** MolViewSpec tree */
-    root: MVSTree,
-    /** Associated metadata */
-    metadata: MVSMetadata,
-}
-
-interface MVSMetadata {
-    /** Version of the spec used to write this tree */
-    version: string,
-    /** Name of this view */
+/** Top-level metadata for a MVS file (single-state or multi-state). */
+export interface GlobalMetadata {
+    /** Name of this MVSData */
     title?: string,
     /** Detailed description of this view */
     description?: string,
-    /** Format of the description */
+    /** Format of `description`. Default is 'markdown'. */
     description_format?: 'markdown' | 'plaintext',
-    /** Timestamp when this view was exported */
+    /** Timestamp when this view was exported. */
     timestamp: string,
+    /** Version of MolViewSpec used to write this file. */
+    version: string,
 }
+export const GlobalMetadata = {
+    create(metadata?: Pick<GlobalMetadata, 'title' | 'description' | 'description_format'>): GlobalMetadata {
+        return {
+            ...metadata,
+            version: `${MVSData.SupportedVersion}`,
+            timestamp: utcNowISO(),
+        };
+    },
+};
+
+/** Metadata for an individual snapshot. */
+export interface SnapshotMetadata {
+    /** Name of this snapshot. */
+    title?: string,
+    /** Detailed description of this snapshot. */
+    description?: string,
+    /** Format of `description`. Default is 'markdown'. */
+    description_format?: 'markdown' | 'plaintext',
+    /** Unique identifier of this state, useful when working with collections of states. */
+    key?: string,
+    /** Timespan for snapshot. */
+    linger_duration_ms: number,
+    /** Timespan for the animation to the next snapshot. Leave empty to skip animations. */
+    transition_duration_ms?: number,
+}
+
+export interface Snapshot {
+    /** Root of the node tree */
+    root: MVSTree,
+    /** Associated metadata */
+    metadata: SnapshotMetadata,
+}
+
+/** MVSData with a single state */
+export interface MVSData_State {
+    kind?: 'single',
+    /** Root of the node tree */
+    root: MVSTree,
+    /** Associated metadata */
+    metadata: GlobalMetadata,
+}
+
+/** MVSData with multiple states (snapshots) */
+export interface MVSData_States {
+    kind: 'multiple',
+    /** Ordered collection of individual states */
+    snapshots: Snapshot[],
+    /** Associated metadata */
+    metadata: GlobalMetadata,
+}
+
+/** Top level of the MolViewSpec (MVS) data format. */
+export type MVSData = MVSData_State | MVSData_States
+
 
 export const MVSData = {
     /** Currently supported major version of MolViewSpec format (e.g. 1 for version '1.0.8') */
@@ -62,15 +108,31 @@ export const MVSData = {
      * If `options.noExtra` is true, presence of any extra node parameters is treated as an issue. */
     validationIssues(mvsData: MVSData, options: { noExtra?: boolean } = {}): string[] | undefined {
         const version = mvsData?.metadata?.version;
-        if (typeof version !== 'string') return [`"version" in MVS must be a string, not ${typeof version}: ${version}`];
-        if (mvsData.root === undefined) return [`"root" missing in MVS`];
-        return treeValidationIssues(MVSTreeSchema, mvsData.root, options);
+        if (typeof version !== 'string') return [`MVSData.metadata.version must be a string, not ${typeof version}: ${version}`];
+        if (mvsData.kind === 'single' || mvsData.kind === undefined) {
+            return snapshotValidationIssues(mvsData, options);
+        } else if (mvsData.kind === 'multiple') {
+            if (mvsData.snapshots === undefined) return [`"snapshots" missing in MVS`];
+            const issues: string[] = [];
+            for (const snapshot of mvsData.snapshots) { // would use .flatMap if it didn't work in a completely unpredictable way
+                const snapshotIssues = snapshotValidationIssues(snapshot, options);
+                if (snapshotIssues) issues.push(...snapshotIssues);
+            }
+            if (issues.length > 0) return issues;
+            else return undefined;
+        } else {
+            return [`MVSData.kind must be 'single' or 'multiple', not ${mvsData.kind}`];
+        }
     },
 
     /** Return a human-friendly textual representation of `mvsData`. */
     toPrettyString(mvsData: MVSData): string {
+        const type = mvsData.kind === 'multiple' ? 'multiple states' : 'single state';
         const title = mvsData.metadata.title !== undefined ? ` "${mvsData.metadata.title}"` : '';
-        return `MolViewSpec tree${title} (version ${mvsData.metadata.version}, created ${mvsData.metadata.timestamp}):\n${treeToString(mvsData.root)}`;
+        const trees = mvsData.kind === 'multiple' ?
+            mvsData.snapshots.map((s, i) => `[Snapshot #${i}]\n${treeToString(s.root)}`).join('\n')
+            : treeToString(mvsData.root);
+        return `MolViewSpec ${type}${title} (version ${mvsData.metadata.version}, created ${mvsData.metadata.timestamp}):\n${trees}`;
     },
 
     /** Create a new MolViewSpec builder containing only a root node. Example of MVS builder usage:
@@ -86,6 +148,15 @@ export const MVSData = {
     createBuilder(): Root {
         return createMVSBuilder();
     },
+
+    /** Create a multi-state MVS data from a list of snapshots. */
+    createMultistate(snapshots: Snapshot[], metadata?: Pick<GlobalMetadata, 'title' | 'description' | 'description_format'>): MVSData_States {
+        return {
+            kind: 'multiple',
+            snapshots: [...snapshots],
+            metadata: GlobalMetadata.create(metadata),
+        };
+    },
 };
 
 
@@ -95,4 +166,14 @@ function majorVersion(semanticVersion: string | number): number | undefined {
     if (typeof semanticVersion === 'number') return Math.floor(semanticVersion);
     console.error(`Version should be a string, not ${typeof semanticVersion}: ${semanticVersion}`);
     return undefined;
+}
+
+function snapshotValidationIssues(snapshot: MVSData_State | Snapshot, options: { noExtra?: boolean } = {}): string[] | undefined {
+    if (snapshot.root === undefined) return [`"root" missing in snapshot`];
+    return treeValidationIssues(MVSTreeSchema, snapshot.root, options);
+}
+
+/** Return the current universal time, in ISO format, e.g. '2023-11-24T10:45:49.873Z' */
+function utcNowISO(): string {
+    return new Date().toISOString();
 }